Fix E0308 errors in mirrorcoat, perishsong, phantomforce, pikapapow
- mirrorcoat.rs: Fix get_at_slot to pass Option<&str>, fix add_volatile signature
- perishsong.rs: Fix run_event result checks - use Some(0) and is_none() instead of EventResult
- phantomforce.rs: Fix run_event result check - use Some(0) instead of EventResult
- pikapapow.rs: Cast bp to i32 for EventResult::Number

+++ src/data/move_callbacks/pikapapow.rs
@@ -30,6 +30,6 @@ pub fn base_power_callback(battle: &mut Battle, pokemon_pos: (usize, usize), tar
     // (debug is typically not needed in Rust implementation)
 
     // return bp;
-    EventResult::Number(bp)
+    EventResult::Number(bp as i32)
 }
